requirements-backend.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
supabase==2.6.0
pydantic==2.9.0
pydantic-settings==2.5.0
spacy==3.7.6
openai==1.47.0
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.32.0
pdfplumber==0.11.0
python-docx==1.1.2
pytesseract==0.3.13
pdf2image==1.17.0
playwright==1.47.0
aiofiles==24.1.0
python-dotenv==1.0.1
